finish comment component

+++ tags/cloud_blog/src/main/java/com/zhela/cloudblog/service/ourservice/impl/CommentServiceImpl.java
@@ -1,5 +1,11 @@
 package com.zhela.cloudblog.service.ourservice.impl;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 import java.util.List;
 
 import org.apache.commons.lang.StringUtils;
@@ -18,6 +24,7 @@ import com.zhela.cloudblog.service.tweet.provider.impl.ProviderTweetServiceSINA;
 import com.zhela.cloudblog.util.SystemConfig;
 
 public class CommentServiceImpl implements CommentService {
+	
 
 	CommentDAO commentDAO;
 	ServiceManage serviceManage;
@@ -29,7 +36,7 @@ public class CommentServiceImpl implements CommentService {
 	public RESTServiceCommentList loadComment(String code,String web_url, int page,
 			int size,String order) throws Exception {
 		if(!serviceManage.isAllowed(code, web_url)){
-			throw new Exception("你没有权限");
+			throw new Exception("你没有注册过该服务或者你已经停止了该服务");
 		}
 		if(page<=0){
 			page = 1;
@@ -60,6 +67,42 @@ public class CommentServiceImpl implements CommentService {
 		restcomment.setComment(_comments);
 		return restcomment;
 	}
+	
+	@Override
+	public RESTServiceCommentList loadComment(String code, int page,
+			int size,String order,String account) throws Exception {
+		if(!serviceManage.isAllowedAccount(code, account)){
+			throw new Exception("你没有权限查看该统计");
+		}
+		if(page<=0){
+			page = 1;
+		}
+		RESTServiceCommentList restcomment = new RESTServiceCommentList();
+		CommentExample example = new CommentExample();
+		Criteria criteria = example.createCriteria();
+		criteria.andServiceBindCodeEqualTo(code);
+		example.setPage(page, size);
+		int count = commentDAO.countByExample(example);
+		restcomment.setCount(count);
+		if(count>page*size){
+			if(count>size){
+				restcomment.setNextPage(page+1);
+			}
+		}
+		if(page>1){
+			restcomment.setLastPage(page-1);
+		}
+		example.setOrderByClause("web_url,id "+order);
+		List<Comment> comments = commentDAO.selectByExampleWithBLOBs(example);
+		java.util.List<RESTServiceComment> _comments = new java.util.ArrayList<RESTServiceComment>();
+		RESTServiceComment rest;
+		for(Comment _comment:comments){
+			rest = getRESTFromComment(_comment);
+			_comments.add(rest);
+		}
+		restcomment.setComment(_comments);
+		return restcomment;
+	}
 
 	private RESTServiceComment getRESTFromComment(Comment _comment){
 		RESTServiceComment rest = new RESTServiceComment();
@@ -68,6 +111,9 @@ public class CommentServiceImpl implements CommentService {
 		rest.setUpdate_time(_comment.getUpdateTime());
 		rest.setUserheader(SystemConfig.domain+_comment.getUserHeader());
 		rest.setUsername(_comment.getUserName());
+		rest.setWeb_url(_comment.getWebUrl());
+		rest.setService_code(_comment.getServiceBindCode());
+		rest.setIp(_comment.getUserIp());
 		return rest;
 	}
 	
@@ -109,13 +155,65 @@ public class CommentServiceImpl implements CommentService {
 	}
 
 	@Override
-	public RESTServiceComment deleteComment(long id) {
+	public RESTServiceComment deleteComment(String account,long id) {
 		Comment comment = commentDAO.selectByPrimaryKey(id);
 		if(comment!=null){
-			commentDAO.deleteByPrimaryKey(id);
-			return getRESTFromComment(comment);
+			if(serviceManage.isAllowedAccount(comment.getServiceBindCode(), account)){
+				commentDAO.deleteByPrimaryKey(id);
+				return getRESTFromComment(comment);
+			}
 		}
 		return null;
+	}
+
+	@Override
+	public int loadCommentCount(String account,String code)throws Exception {
+		if(!serviceManage.isAllowedAccount(code, account)){
+			throw new Exception("你没有权限查看该统计");
+		}
+		CommentExample example =  new CommentExample();
+		example.createCriteria().andServiceBindCodeEqualTo(code);
+		return commentDAO.countByExample(example);
+	}
+	
+	private final static String comment_template = "/zl_js/service/comment/template/zl_comment_template.js";
+	@Override
+	public void createCommentFile(String code, String path) throws Exception {
+		File file = new File(path+comment_template);
+		Long filelength = file.length();
+		byte[] filecontent = new byte[filelength.intValue()];
+		try {
+			FileInputStream in = new FileInputStream(file);
+			in.read(filecontent);
+			in.close();
+		} catch (FileNotFoundException e) {
+			throw new Exception("生产文件错误，模板文件没找到");
+		} catch (IOException e) {
+			throw new Exception("生产文件错误，模板文件读取错误");
+		}
+		try {
+			String jstext = new String(filecontent, "utf-8");
+			jstext = jstext.replaceAll("zlIDzl", code);
+			file = new File(path+"/zl_js/service/comment/"+code+".js");
+			FileOutputStream out = new FileOutputStream(file);
+			out.write(jstext.getBytes("utf-8"));
+			out.flush();
+			out.close();
+		} catch (UnsupportedEncodingException e) {
+		}
+	}
+
+	@Override
+	public void deleteCommentFile(String path) throws Exception {
+		File file = new File(path);
+		file.delete();
+	}
+	
+	@Override
+	public void deleteCommentbyCode(String code) throws Exception {
+		CommentExample example =  new CommentExample();
+		example.createCriteria().andServiceBindCodeEqualTo(code);
+		commentDAO.deleteByExample(example);
 	}
 	
 	
@@ -163,5 +261,6 @@ public class CommentServiceImpl implements CommentService {
 		}
 		
 	}
+
 	
 }
